Added functionality to manage shipping methods
This is 1st subsection of assignment 1. (1.A)
Any admin should be able to add shipping methods,
add for multiple countries or delete for a country
via rails console.
Only add_for_countries method is sufficient to the
way the assignment requirement.
The rest (add and delete_for_country) of the methods
can be used as helpers.

Usage from rails console -
1. ShippingMethod.add_for_countries("fedex", {"IN": 3, "US": 6})
will add the `fedex` delivery method for India with 3 delivery days
and United states with 6 delivery days. Use this method to add
same delivery method to as many countries as needed.

2. ShippingMethod.add({name: "fedex", country: "IN",
   delivery_time_in_days: 3})
will add the `fedex` delivery method for India with 3 delivery days.
Use this method to add delivery method for a single country.

Note: Trying to add the same country again will result in an error.
As, a country can have only one delivery method.

3. ShippingMethod.delete_for_country("IN")
will delete the record for that country. Will raise error
if record doesn't exist.

All 3 public methods have unit tests in shipping_method_spec.rb