[UI] Added stimulus controller and its trigger point
On selecting country, stimulus will fire a requestjs request to the
server. Country value, tag ids for Submit button, new paragraph tag
for showing delivery info and turbo_frame_target are collected and
sent as get request to the server. This request will expect a
turbo_stream response.
This will be required for Assignment 1B