Added schema and migrations for ERP data
Rationale behind no foreign key with orders -
ERP can be an external system and may need to be abstracted
out from the database. Foreign keys bind multiple tables together
and this seems like a use case where we may have to keep things
independent. This is done at the cost of integrity which is fine to
loose. If an order is being deleted, keeping the existing erp record
doesn't harm us.

Rationale behind individual order_id and entity_id indexes -
ERP system administrators may query the data with either order or
entity. Hence this takes care of the uniqueness and allow fast search
based on the search field.